requirements.txt
django-rest-swagger==2.2.0
django-cors-headers==3.2.0
djangorestframework==3.10.3
mysql-connector-python==8.0.18
pdfkit==0.6.1
numpy==1.17.4
pandas==0.25.3
Jinja2==2.10.3
dask==2.8.0
astroid==2.3.0
Django==2.2.5
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.4.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.2.0
virtualenv==16.7.5
wrapt==1.11.2
